Sprint 12. Start to refactoring user info download sequence.
