+++ ImageFeed/Models/OAuthTokenResponse.swift
@@ -1,10 +1,3 @@
-//
-//  OAuthTokenResponse.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 12/19/22.
-//
-
 import Foundation
 
 fileprivate enum CodingKeys: String, CodingKey {
