+++ ImageFeed/Models/Photo.swift
@@ -1,10 +1,3 @@
-//
-//  Model.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 1/23/23.
-//
-
 import Foundation
 
 struct Photo {
