+++ ImageFeed/Services/ProfileService/ProfileService.swift
@@ -1,10 +1,3 @@
-//
-//  ProfileService.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 1/3/23.
-//
-
 import Foundation
 
 final class ProfileService {
