+++ ImageFeed/Services/UIBlockingProgressHUD/UIBlockingProgressHUD.swift
@@ -1,10 +1,3 @@
-//
-//  UIBlockingProgressHUD.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 1/3/23.
-//
-
 import UIKit
 import ProgressHUD
