+++ ImageFeed/Services/URLSessionExtension/URLSessionExtension.swift
@@ -1,15 +1,6 @@
-//
-//  URLSessionExtension.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 1/9/23.
-//
-
 import Foundation
 
 extension URLSession {
-    
-    //MARK: - Generic method
     func objectTask<T: Decodable> (
         for request: URLRequest,
         completion: @escaping (Result<T, Error>) -> Void
@@ -27,14 +18,11 @@ extension URLSession {
                     return
                 }
                 guard let data = data else { return }
-                print("\n↔️\ndata; \(String(data: data, encoding: .utf8))")
                 do {
                     let decodedObject = try JSONDecoder().decode(T.self, from: data)
                     completion(.success(decodedObject))
-                    print("\n✅🟢\ndecodedObject = \(decodedObject)\nType is: \(type(of: decodedObject))\n")
                 } catch {
                     completion(.failure(NetworkError.decodeError))
-                    print("\n‼️❌⭕️\nNetworkError.decodeError = \(NetworkError.decodeError)")
                 }
             }
         })
