+++ ImageFeed/ViewControllers/AuthView/AuhtViewController.swift
@@ -1,10 +1,3 @@
-//
-//  AuhtViewController.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 12/16/22.
-//
-
 import UIKit
 
 protocol AuthViewControllerDelegate: AnyObject {
