+++ ImageFeed/ViewControllers/AuthView/Constants.swift
@@ -1,10 +1,3 @@
-//
-//  Constants.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 12/16/22.
-//
-
 import Foundation
 
 enum Constants {
