+++ ImageFeed/ViewControllers/TabBarController/TabBarController.swift
@@ -1,10 +1,3 @@
-//
-//  TabBarController.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 1/19/23.
-//
-
 import Foundation
 import UIKit
