+++ ImageFeed/ViewControllers/WebView/WebViewViewController.swift
@@ -1,10 +1,3 @@
-//
-//  WebViewViewController.swift
-//  ImageFeed
-//
-//  Created by Роман Бойко on 12/16/22.
-//
-
 import WebKit
 import UIKit
